+++ lib/create_climb/create_climb_event.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 abstract class CreateClimbEvent {}
 
 class GradeSelectedEvent extends CreateClimbEvent {
@@ -16,6 +18,13 @@ class CategoriesUpdatedEvent extends CreateClimbEvent {
   CategoriesUpdatedEvent(this.selectedCategories);
 
   final List<String> selectedCategories;
+}
+
+class ImageFileUpdatedEvent extends CreateClimbEvent {
+  ImageFileUpdatedEvent(this.deleteImage, this.imageFile);
+
+  final bool deleteImage;
+  final File imageFile;
 }
 
 class CreateClimbErrorEvent extends CreateClimbEvent {}