+++ CHANGELOG.md
@@ -16,5 +16,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Location and climb editing
 - Location, climb, attempt and category deletion
 - Climb filtering and archiving
-- Image addition and deletion for locations
+- Image addition and deletion for locations and climbs
 - App icon